requirements.txt
streamlit
pandas
numpy
opencv-python
pillow
requests
ultralytics
python-dotenv
pydantic
